+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixes
 
+* When deleting a draft document, we remove related reverse IDs of documents having a relation to the deleted one.
 * Fix publishing or moving published page after a draft page on the same tree level to work as expected.
 
 ## 3.62.0 (2024-01-25)